+++ clients/admin-ui/src/features/common/hooks/useInterzoneNav.ts
@@ -1,3 +1,4 @@
+import { SYSTEM_ROUTE } from "~/constants";
 import { useFeatures } from "~/features/common/features";
 import { resolveLink } from "~/features/common/nav/zone-config";
 
@@ -13,7 +14,7 @@ export const useInterzoneNav = () => {
    * Often we need to either go to the systems page or the datamap page
    * depending on if this is an open source or plus instance
    */
-  const systemOrDatamapRoute = features.plus ? datamapRoute.href : "/system";
+  const systemOrDatamapRoute = features.plus ? datamapRoute.href : SYSTEM_ROUTE;
 
   return { systemOrDatamapRoute };
 };